requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
geopy==2.4.1
python-socketio==5.10.0
eventlet==0.33.3
pytest==7.4.3
pytest-flask==1.3.0
Flask-Caching==2.1.0
Authlib==1.3.0
